New version, 1.0.0.11.
This version introduce the so wanted feature that facilitate the blocking of CIDR ip address subnet, in case user want to block some CIDRs using the Windows Firewall. It is required to run the application as administrator for that feature, so if you select to use that feature, you will need to restart the application and there will be a prompt for it.